+++ src/config/index.js
@@ -1,3 +1,4 @@
 module.exports = {
     EVM_NETWORKS_API: 'https://raw.githubusercontent.com/getsafle/multichain-data/main/chains.json',
+    FALLBACK_EVM_NETWORKS_API: 'https://chain-data-safle.netlify.app/chains.json',
 }