+++ src/index.js
@@ -2,25 +2,35 @@ const config = require('./config');
 const helper = require('./utils/helper');
 
 async function getNetworks() {
+    let output;
+
     const { response, error } = await helper.getRequest(config.EVM_NETWORKS_API);
 
     if (error) {
-        return { error };
+        const { response } = await helper.getRequest({ url: config.FALLBACK_EVM_NETWORKS_API });
+
+        output = response;
+    } else {
+        output = response;
     }
 
-    let networks = [];
-
-    return { response };
+    return { output };
 }
 
 async function getChainDetails(chain) {
+    let output;
+
     const { response, error } = await helper.getRequest(config.EVM_NETWORKS_API);
 
     if (error) {
-        return { error };
+        const { response } = await helper.getRequest({ url: config.FALLBACK_EVM_NETWORKS_API });
+
+        output = response;
+    } else {
+        output = response;
     }
 
-    let chainData = response[chain];
+    let chainData = output[chain];
 
     if (chainData === undefined) {
         return { error: 'This chain or network doesn\'t exist' };